Location picking with component
