+++ widgets/events/components/index.ts
@@ -1,1 +1,2 @@
 export { EventsList, type WPEvent } from './events-list';
+export { LocationPicker } from './location-picker';
