+++ widgets/events/components/location-picker/index.ts
@@ -1,0 +1,1 @@
+export { LocationPicker } from './location-picker';
