+++ widgets/events/style.module.css
@@ -1,13 +1,3 @@
-.locationPicker {
-	margin-block-end: var(--wpds-dimension-padding-md);
-	max-width: calc(var(--wpds-dimension-base) * 100);
-}
-
-.locationInput {
-	flex: 1;
-	min-inline-size: calc(var(--wpds-dimension-base) * 30);
-}
-
 .eventNone {
 	padding-block: var(--wpds-dimension-padding-sm);
 	color: var(--wpds-color-fg-content-neutral-weak);
